use film length as int instead of doubles for state

+++ src/main/kotlin/mediathek/gui/tabs/tab_film/filter/DialogFilterView.kt
@@ -50,8 +50,8 @@ class DialogFilterView(
         themaSelectionRenderer(state.thema)
 
         filmLengthRangeSlider.withValueIsAdjusting {
-            filmLengthRangeSlider.lowValue = state.filmLengthMin.toInt()
-            filmLengthRangeSlider.highValue = state.filmLengthMax.toInt()
+            filmLengthRangeSlider.lowValue = state.filmLengthMin
+            filmLengthRangeSlider.highValue = state.filmLengthMax
         }
         minFilmLengthValueLabel.text = filmLengthRangeSlider.lowValue.toString()
         maxFilmLengthValueLabel.text = filmLengthRangeSlider.highValueText
