+++ src/test/kotlin/mediathek/gui/tabs/tab_film/filter/FilterSwitchReloadDecisionTest.kt
@@ -83,8 +83,8 @@ internal class FilterSwitchReloadDecisionTest {
             dontShowDuplicates = false,
             checkedChannels = emptySet(),
             thema = "",
-            filmLengthMin = 0.0,
-            filmLengthMax = FilmLengthSlider.UNLIMITED_VALUE.toDouble(),
+            filmLengthMin = 0,
+            filmLengthMax = FilmLengthSlider.UNLIMITED_VALUE,
             zeitraum = zeitraum
         )
     }
